+++ core/tasks/handler/worker.go
@@ -11,7 +11,6 @@ import (
 	"github.com/nyaruka/gocommon/analytics"
 	"github.com/nyaruka/gocommon/dbutil"
 	"github.com/nyaruka/gocommon/urns"
-	"github.com/nyaruka/goflow/assets"
 	"github.com/nyaruka/goflow/excellent/types"
 	"github.com/nyaruka/goflow/flows"
 	"github.com/nyaruka/goflow/flows/engine"
@@ -21,6 +20,7 @@ import (
 	"github.com/nyaruka/goflow/utils"
 	"github.com/nyaruka/mailroom"
 	"github.com/nyaruka/mailroom/core/models"
+	"github.com/nyaruka/mailroom/core/msgio"
 	"github.com/nyaruka/mailroom/core/queue"
 	"github.com/nyaruka/mailroom/core/runner"
 	"github.com/nyaruka/mailroom/runtime"
@@ -478,28 +478,48 @@ func handleStopEvent(ctx context.Context, rt *runtime.Runtime, event *StopEvent)
 func handleMsgEvent(ctx context.Context, rt *runtime.Runtime, event *MsgEvent) error {
 	oa, err := models.GetOrgAssets(ctx, rt, event.OrgID)
 	if err != nil {
-		return errors.Wrapf(err, "error loading org")
+		return errors.Wrap(err, "error loading org")
+	}
+
+	// load the channel for this message
+	channel := oa.ChannelByID(event.ChannelID)
+
+	// fetch the attachments on the message (i.e. ask courier to fetch them)
+	attachments := make([]utils.Attachment, 0, len(event.Attachments))
+	logUUIDs := make([]models.ChannelLogUUID, 0, len(event.Attachments))
+
+	// no channel, no attachments
+	if channel != nil {
+		for _, attURL := range event.Attachments {
+			// if courier has already fetched this attachment, use it as is
+			if utils.Attachment(attURL).ContentType() != "" {
+				attachments = append(attachments, utils.Attachment(attURL))
+			} else {
+				attachment, logUUID, err := msgio.FetchAttachment(ctx, rt, channel, attURL)
+				if err != nil {
+					return errors.Wrapf(err, "error fetching attachment '%s'", attURL)
+				}
+
+				attachments = append(attachments, attachment)
+				logUUIDs = append(logUUIDs, logUUID)
+			}
+		}
 	}
 
 	// load our contact
-	contacts, err := models.LoadContacts(ctx, rt.ReadonlyDB, oa, []models.ContactID{event.ContactID})
+	modelContact, err := models.LoadContact(ctx, rt.ReadonlyDB, oa, event.ContactID)
 	if err != nil {
 		return errors.Wrapf(err, "error loading contact")
 	}
 
-	// contact has been deleted, ignore this message but mark it as handled
-	if len(contacts) == 0 {
-		err := models.UpdateMessage(ctx, rt.DB, event.MsgID, models.MsgStatusHandled, models.VisibilityArchived, models.MsgTypeInbox, models.NilFlowID)
+	// contact has been deleted, or is blocked, or channel no longer exists, ignore this message but mark it as handled
+	if modelContact == nil || modelContact.Status() == models.ContactStatusBlocked || channel == nil {
+		err := models.UpdateMessage(ctx, rt.DB, event.MsgID, models.MsgStatusHandled, models.VisibilityArchived, models.MsgTypeInbox, models.NilFlowID, attachments, logUUIDs)
 		if err != nil {
 			return errors.Wrapf(err, "error updating message for deleted contact")
 		}
 		return nil
 	}
-
-	modelContact := contacts[0]
-
-	// load the channel for this message
-	channel := oa.ChannelByUUID(event.ChannelUUID)
 
 	// if we have URNs make sure the message URN is our highest priority (this is usually a noop)
 	if len(modelContact.URNs()) > 0 {
@@ -507,15 +527,6 @@ func handleMsgEvent(ctx context.Context, rt *runtime.Runtime, event *MsgEvent) e
 		if err != nil {
 			return errors.Wrapf(err, "error changing primary URN")
 		}
-	}
-
-	// if this channel is no longer active or this contact is blocked, ignore this message (mark it as handled)
-	if channel == nil || modelContact.Status() == models.ContactStatusBlocked {
-		err := models.UpdateMessage(ctx, rt.DB, event.MsgID, models.MsgStatusHandled, models.VisibilityArchived, models.MsgTypeInbox, models.NilFlowID)
-		if err != nil {
-			return errors.Wrapf(err, "error marking blocked or nil channel message as handled")
-		}
-		return nil
 	}
 
 	// stopped contact? they are unstopped if they send us an incoming message
@@ -549,7 +560,7 @@ func handleMsgEvent(ctx context.Context, rt *runtime.Runtime, event *MsgEvent) e
 		return errors.Wrapf(err, "unable to look up open tickets for contact")
 	}
 	for _, ticket := range tickets {
-		ticket.ForwardIncoming(ctx, rt, oa, event.MsgUUID, event.Text, event.Attachments)
+		ticket.ForwardIncoming(ctx, rt, oa, event.MsgUUID, event.Text, attachments)
 	}
 
 	// find any matching triggers
@@ -577,7 +588,7 @@ func handleMsgEvent(ctx context.Context, rt *runtime.Runtime, event *MsgEvent) e
 		}
 	}
 
-	msgIn := flows.NewMsgIn(event.MsgUUID, event.URN, channel.ChannelReference(), event.Text, event.Attachments)
+	msgIn := flows.NewMsgIn(event.MsgUUID, event.URN, channel.ChannelReference(), event.Text, attachments)
 	msgIn.SetExternalID(string(event.MsgExternalID))
 	msgIn.SetID(event.MsgID)
 
@@ -589,7 +600,7 @@ func handleMsgEvent(ctx context.Context, rt *runtime.Runtime, event *MsgEvent) e
 		}
 		sessions[0].SetIncomingMsg(event.MsgID, event.MsgExternalID)
 
-		return markMsgHandled(ctx, tx, contact, msgIn, flow, tickets)
+		return markMsgHandled(ctx, tx, contact, msgIn, flow, attachments, tickets, logUUIDs)
 	}
 
 	// we found a trigger and their session is nil or doesn't ignore keywords
@@ -606,7 +617,7 @@ func handleMsgEvent(ctx context.Context, rt *runtime.Runtime, event *MsgEvent) e
 			// if this is an IVR flow, we need to trigger that start (which happens in a different queue)
 			if flow.FlowType() == models.FlowTypeVoice {
 				ivrMsgHook := func(ctx context.Context, tx *sqlx.Tx) error {
-					return markMsgHandled(ctx, tx, contact, msgIn, flow, tickets)
+					return markMsgHandled(ctx, tx, contact, msgIn, flow, attachments, tickets, logUUIDs)
 				}
 				err = runner.TriggerIVRFlow(ctx, rt, oa.OrgID(), flow.ID(), []models.ContactID{modelContact.ID()}, ivrMsgHook)
 				if err != nil {
@@ -636,7 +647,7 @@ func handleMsgEvent(ctx context.Context, rt *runtime.Runtime, event *MsgEvent) e
 	}
 
 	// this message didn't trigger and new sessions or resume any existing ones, so handle as inbox
-	err = handleAsInbox(ctx, rt, oa, contact, msgIn, tickets)
+	err = handleAsInbox(ctx, rt, oa, contact, msgIn, attachments, logUUIDs, tickets)
 	if err != nil {
 		return errors.Wrapf(err, "error handling inbox message")
 	}
@@ -740,7 +751,7 @@ func handleTicketEvent(ctx context.Context, rt *runtime.Runtime, event *models.T
 }
 
 // handles a message as an inbox message
-func handleAsInbox(ctx context.Context, rt *runtime.Runtime, oa *models.OrgAssets, contact *flows.Contact, msg *flows.MsgIn, tickets []*models.Ticket) error {
+func handleAsInbox(ctx context.Context, rt *runtime.Runtime, oa *models.OrgAssets, contact *flows.Contact, msg *flows.MsgIn, attachments []utils.Attachment, logUUIDs []models.ChannelLogUUID, tickets []*models.Ticket) error {
 	// usually last_seen_on is updated by handling the msg_received event in the engine sprint, but since this is an inbox
 	// message we manually create that event and handle it
 	msgEvent := events.NewMsgReceived(msg)
@@ -752,11 +763,11 @@ func handleAsInbox(ctx context.Context, rt *runtime.Runtime, oa *models.OrgAsset
 		return errors.Wrap(err, "error handling inbox message events")
 	}
 
-	return markMsgHandled(ctx, rt.DB, contact, msg, nil, tickets)
+	return markMsgHandled(ctx, rt.DB, contact, msg, nil, attachments, tickets, logUUIDs)
 }
 
 // utility to mark as message as handled and update any open contact tickets
-func markMsgHandled(ctx context.Context, db models.Queryer, contact *flows.Contact, msg *flows.MsgIn, flow *models.Flow, tickets []*models.Ticket) error {
+func markMsgHandled(ctx context.Context, db models.Queryer, contact *flows.Contact, msg *flows.MsgIn, flow *models.Flow, attachments []utils.Attachment, tickets []*models.Ticket, logUUIDs []models.ChannelLogUUID) error {
 	msgType := models.MsgTypeInbox
 	flowID := models.NilFlowID
 	if flow != nil {
@@ -764,7 +775,7 @@ func markMsgHandled(ctx context.Context, db models.Queryer, contact *flows.Conta
 		flowID = flow.ID()
 	}
 
-	err := models.UpdateMessage(ctx, db, msg.ID(), models.MsgStatusHandled, models.VisibilityVisible, msgType, flowID)
+	err := models.UpdateMessage(ctx, db, msg.ID(), models.MsgStatusHandled, models.VisibilityVisible, msgType, flowID, attachments, logUUIDs)
 	if err != nil {
 		return errors.Wrapf(err, "error marking message as handled")
 	}
@@ -791,19 +802,18 @@ type TimedEvent struct {
 }
 
 type MsgEvent struct {
-	ContactID     models.ContactID   `json:"contact_id"`
-	OrgID         models.OrgID       `json:"org_id"`
-	ChannelUUID   assets.ChannelUUID `json:"channel_uuid"`
-	ChannelType   models.ChannelType `json:"channel_type"`
-	MsgID         flows.MsgID        `json:"msg_id"`
-	MsgUUID       flows.MsgUUID      `json:"msg_uuid"`
-	MsgExternalID null.String        `json:"msg_external_id"`
-	URN           urns.URN           `json:"urn"`
-	URNID         models.URNID       `json:"urn_id"`
-	Text          string             `json:"text"`
-	Attachments   []utils.Attachment `json:"attachments"`
-	NewContact    bool               `json:"new_contact"`
-	CreatedOn     time.Time          `json:"created_on"`
+	ContactID     models.ContactID `json:"contact_id"`
+	OrgID         models.OrgID     `json:"org_id"`
+	ChannelID     models.ChannelID `json:"channel_id"`
+	MsgID         flows.MsgID      `json:"msg_id"`
+	MsgUUID       flows.MsgUUID    `json:"msg_uuid"`
+	MsgExternalID null.String      `json:"msg_external_id"`
+	URN           urns.URN         `json:"urn"`
+	URNID         models.URNID     `json:"urn_id"`
+	Text          string           `json:"text"`
+	Attachments   []string         `json:"attachments"`
+	NewContact    bool             `json:"new_contact"`
+	CreatedOn     time.Time        `json:"created_on"`
 }
 
 type StopEvent struct {