+++ CHANGELOG.md
@@ -1,3 +1,18 @@
+v7.5.18
+----------
+ * Move interrupt_channel task into tasks/interrupts
+ * Tweak msg retries to ignore deleted channels
+ * Remove no longer used channels option from interrupt_sessions task
+
+v7.5.17
+----------
+ * Update to latest goflow
+ * Add interrupt channel task
+
+v7.5.16
+----------
+ * If starting a message flow from an ivr trigger, send reject response
+
 v7.5.15
 ----------
  * Update to latest goflow which changes expirations for dial waits